+++ apps/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web
 
+## 0.6.11
+
+### Patch Changes
+
+- @shallot-ui/prismjs@18.0.1
+
 ## 0.6.10
 
 ### Patch Changes